move filters into submodule
