+++ discovery_engine_core/core/src/stack/ops/breaking.rs
@@ -26,8 +26,7 @@ use crate::{
     document::{Document, HistoricDocument},
     engine::{EndpointConfig, GenericError},
     stack::{
-        filters::{ArticleFilter, CommonFilter},
-        semantic_filters::{filter_semantically, SemanticFilterConfig},
+        filters::{filter_semantically, ArticleFilter, CommonFilter, SemanticFilterConfig},
         Id,
     },
 };
